refactor: purge blog components
and move remaining folders from /blog to /components
