--- src/blog/components/hero/blogHero.module.css
@@ -1,89 +1,0 @@
-.hero {
-  width: 100%;
-  max-width: 1200px;
-  margin: auto;
-  display: flex;
-  flex-direction: column;
-  gap: 3rem;
-  align-self: stretch;
-  padding-top: 7.5rem;
-  padding-bottom: 2.5rem;
-}
-
-.sidePadding {
-  @media (max-width: 640px) {
-    padding-left: 3rem;
-    padding-right: 3rem;
-  }
-
-  @media (max-width: 1400px) {
-    padding-left: 2rem;
-    padding-right: 2rem;
-  }
-
-  @media (min-width: 1400px) {
-    padding-left: 0;
-    padding-right: 0;
-  }
-}
-
-.tabList {
-  composes: sidePadding;
-  list-style: none;
-  margin: 0;
-  padding-top: 6px;
-  padding-bottom: 6px;
-  display: inline-flex;
-  align-items: center;
-  flex-direction: row;
-  gap: 1rem;
-  white-space: nowrap;
-  overflow: auto;
-  scrollbar-width: none; /* Firefox */
-  /* Hide scrollbar for Chrome, Safari and Opera */
-  &::-webkit-scrollbar {
-    display: none;
-  }
-}
-
-.tab {
-  all: unset;
-  cursor: pointer;
-  display: inline-block;
-  padding: 0.375rem 0.75rem;
-  border-radius: 1.5rem;
-  background-color: var(--primary-white);
-  color: var(--primary-black);
-  font-family: var(--font-britti-sans);
-  font-size: 1rem;
-  font-weight: 400;
-  white-space: nowrap;
-
-  @media (min-width: 1024px) {
-    font-size: 1.25rem;
-    line-height: 140%;
-    padding: 0.5rem 1.5rem;
-  }
-
-  &:hover {
-    background-color: var(--secondary-off-white1);
-  }
-
-  &:active {
-    background-color: var(--secondary-off-white2);
-  }
-
-  &:focus {
-    outline: 2px solid var(--focus-color);
-    border-radius: 1.5rem;
-  }
-}
-
-.selected {
-  background-color: var(--primary-black);
-  color: var(--primary-white-bright);
-  &:hover {
-    background-color: var(--primary-black);
-    color: var(--primary-white-bright);
-  }
-}
