--- src/blog/components/loadingNews/loadingNews.module.css
@@ -1,48 +1,0 @@
-.wrapper {
-  height: 38rem;
-  background: var(--primary-white, #faf8f5);
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-  align-items: flex-start;
-}
-
-.loading {
-  display: flex;
-  padding: 5rem 2rem 30rem 2rem;
-  flex-direction: row;
-  align-items: flex-start;
-  gap: 1.5rem;
-  align-self: stretch;
-
-  @media (min-width: 640px) {
-    padding: 5rem 3rem 30rem 3rem;
-  }
-
-  @media (min-width: 1024px) {
-    padding: 5rem 8.75rem 30rem 8.75rem;
-  }
-
-  &::before {
-    content: "";
-    display: block;
-    width: 1.5rem;
-    height: 1.5rem;
-    -webkit-mask: url("/_assets/spinner.svg") no-repeat 50% 50%;
-    mask: url("/_assets/spinner.svg") no-repeat 50% 50%;
-    background-color: var(--primary-black);
-    animation: rotateSpinner 750ms linear infinite;
-  }
-}
-
-@keyframes rotateSpinner {
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-@-webkit-keyframes rotateSpinner {
-  to {
-    transform: rotate(360deg);
-  }
-}
