--- src/blog/components/postCard/postCard.module.css
@@ -1,51 +1,0 @@
-.cardWrapper {
-  box-sizing: border-box;
-  display: flex;
-}
-
-.card {
-  padding: 1rem;
-  border-radius: 0.75rem;
-  border: 1px solid var(--secondary-yellow-1, #f5f3e5);
-  cursor: pointer;
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: flex-start;
-  gap: 0.5rem;
-  text-decoration: none;
-  text-wrap: wrap;
-  color: var(--primary-black);
-  font-family: var(--font-britti-sans);
-  font-size: 1.25rem;
-  font-weight: 400;
-  line-height: 140%;
-  flex-grow: 1;
-
-  &:hover {
-    background-color: var(--primary-bg-dark);
-  }
-
-  &:active {
-    background-color: var(--secondary-off-white2);
-  }
-
-  @media (min-width: 1024px) {
-    font-size: 1.5rem;
-  }
-}
-
-.image {
-  width: 100%;
-  height: 12.5rem;
-  border-radius: 0.5rem;
-  overflow: hidden;
-
-  @media (min-width: 1024px) {
-    height: 15rem;
-  }
-
-  @media (min-width: 1400px) {
-    width: 100%;
-  }
-}
