--- src/blog/components/postPreviewGrid/postPreviewGrid.module.css
@@ -1,59 +1,0 @@
-.wrapper {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  overflow: hidden;
-}
-
-.previewGrid {
-  padding: 5rem 1rem;
-  width: -webkit-fill-available;
-  display: flex;
-  flex-direction: column;
-  gap: 24px;
-
-  @media (min-width: 640px) {
-    padding: 5rem 2rem;
-    max-width: 50rem;
-  }
-
-  @media (min-width: 1024px) {
-    max-width: 75rem;
-    min-width: 870px;
-  }
-}
-
-.list {
-  display: grid;
-  gap: 1rem;
-  width: 100%;
-  margin: 0;
-  padding: 0;
-  list-style: none;
-  grid-template-columns: repeat(
-    auto-fill,
-    minmax(100%, 1fr)
-  ); /* Single column by default */
-
-  @media (min-width: 640px) {
-    gap: 1.5rem;
-    grid-template-columns: repeat(
-      auto-fill,
-      minmax(272px, 1fr)
-    ); /* Up to two columns */
-  }
-
-  @media (min-width: 1024px) {
-    grid-template-columns: repeat(
-      3,
-      minmax(272px, 1fr)
-    ); /* Up to three columns */
-  }
-
-  @media (min-width: 1400px) {
-    grid-template-columns: repeat(
-      3,
-      minmax(381px, 1fr)
-    ); /* Still three columns but wider cards */
-  }
-}
