+++ src/components/compensations/components/salaryCalculator/SalaryCalculator.tsx
