+++ src/components/compensations/components/salaryCalculator/salaryCalculator.module.css
