+++ src/components/compensations/utils/salary.ts
