+++ src/components/informationSection/InformationSection.tsx
