+++ src/components/legal/Legal.tsx
