+++ src/components/legal/LegalPreview.tsx
