+++ src/components/utils/linkTypes.ts
@@ -10,22 +10,6 @@ export const homeLink = {
   },
 };
 
-export const studioLink = {
-  _key: "go-to-studio",
-  _type: "link",
-  linkTitle: "Go to studio",
-  linkType: LinkType.Internal,
-  internalLink: { _ref: "studio" },
-};
-
-export const sharedStudioLink = {
-  _key: "go-to-shared-studio",
-  _type: "link",
-  linkTitle: "Go to shared studio",
-  linkType: LinkType.Internal,
-  internalLink: { _ref: "shared" },
-};
-
 export const sharedCustomerCasesLink = {
   _key: "go-to-shared-studio-customer-cases",
   _type: "link",
