--- src/post/hero/postHero.module.css
@@ -1,15 +1,0 @@
-.hero {
-  width: 100%;
-  max-width: 1200px;
-  margin: auto;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: flex-start;
-  padding: 7.5rem 2rem 2.5rem 2rem;
-  gap: 1.5rem;
-
-  @media (min-width: 1024px) {
-    padding: 7.5rem 2rem 5rem 2.5rem;
-  }
-}
