--- src/post/morePostsGrid/mockdata.ts
@@ -1,161 +1,0 @@
-import placeholder from "src/stories/assets/image-placeholder.png";
-import { Post } from "studio/lib/interfaces/pages";
-
-export const mockPosts: Post[] = [
-  {
-    _createdAt: "2024-07-24T09:09:02Z",
-    _updatedAt: "2024-07-24T10:00:00Z",
-    _rev: "v4cJiHBx1xl8CERIziAN7Z",
-    _type: "blogPosts",
-    _id: "0f4dd8bf-9766-4915-9a5b-929e5f08fc87",
-    basicTitle: "Awesome title here",
-    category: "Environment",
-    slug: {
-      _type: "slug",
-      current: "carbon-accounting-explained-16",
-    },
-    lead: {
-      image: {
-        _key: "2052ec0aea39",
-        _type: "image",
-        alt: "Alt text here",
-        src: placeholder,
-      },
-      richText: [
-        {
-          _type: "block",
-          _key: "272e06bbca76",
-          children: [
-            {
-              _key: "dcf62ae1c1f10",
-              _type: "span",
-              marks: [],
-              text: "In a world grappling with the consequences of climate change and environmental degradation, the need for sustainable practices has become more urgent than ever before. As individuals, communities, and businesses, we have a responsibility to make conscious choices that support the health of our planet.",
-            },
-          ],
-          style: "normal",
-        },
-      ],
-    },
-    richText: [
-      {
-        _type: "block",
-        _key: "272e06bbca76",
-        children: [
-          {
-            _key: "dcf62ae1c1f10",
-            _type: "span",
-            marks: [],
-            text: "This is the text content for Carbon Accounting Explained 16.",
-          },
-        ],
-        style: "normal",
-      },
-    ],
-    date: "2024-07-24T09:09:02.000Z",
-  },
-  {
-    _createdAt: "2024-07-24T09:09:02Z",
-    _updatedAt: "2024-07-24T10:00:00Z",
-    _rev: "v4cJiHBx1xl8CERIziAN7Z",
-    _type: "blogPosts",
-    _id: "0f4dd8bf-9766-4915-9a5b-929e5f08fc87",
-    basicTitle: "Another awesome title here",
-    category: "Environment",
-    slug: {
-      _type: "slug",
-      current: "carbon-accounting-explained-16",
-    },
-    lead: {
-      image: {
-        _key: "2052ec0aea39",
-        _type: "image",
-        alt: "Alt text here",
-        src: placeholder,
-      },
-      richText: [
-        {
-          _type: "block",
-          _key: "272e06bbca76",
-          children: [
-            {
-              _key: "dcf62ae1c1f10",
-              _type: "span",
-              marks: [],
-              text: "In a world grappling with the consequences of climate change and environmental degradation, the need for sustainable practices has become more urgent than ever before. As individuals, communities, and businesses, we have a responsibility to make conscious choices that support the health of our planet.",
-            },
-          ],
-          style: "normal",
-        },
-      ],
-    },
-    richText: [
-      {
-        _type: "block",
-        _key: "272e06bbca76",
-        children: [
-          {
-            _key: "dcf62ae1c1f10",
-            _type: "span",
-            marks: [],
-            text: "This is the text content for Carbon Accounting Explained 16.",
-          },
-        ],
-        style: "normal",
-      },
-    ],
-    date: "2024-07-24T09:09:02.000Z",
-  },
-  {
-    _createdAt: "2024-07-24T09:09:02Z",
-    _updatedAt: "2024-07-24T10:00:00Z",
-    _rev: "v4cJiHBx1xl8CERIziAN7Z",
-    _type: "blogPosts",
-    _id: "0f4dd8bf-9766-4915-9a5b-929e5f08fc87",
-    basicTitle: "Cool title",
-    category: "Environment",
-    slug: {
-      _type: "slug",
-      current: "carbon-accounting-explained-16",
-    },
-    lead: {
-      image: {
-        _key: "2052ec0aea39",
-        _type: "image",
-        alt: "Alt text here",
-        src: placeholder,
-      },
-      richText: [
-        {
-          _type: "block",
-          _key: "272e06bbca76",
-          children: [
-            {
-              _key: "dcf62ae1c1f10",
-              _type: "span",
-              marks: [],
-              text: "In a world grappling with the consequences of climate change and environmental degradation, the need for sustainable practices has become more urgent than ever before. As individuals, communities, and businesses, we have a responsibility to make conscious choices that support the health of our planet.",
-            },
-          ],
-          style: "normal",
-        },
-      ],
-    },
-    richText: [
-      {
-        _type: "block",
-        _key: "272e06bbca76",
-        children: [
-          {
-            _key: "dcf62ae1c1f10",
-            _type: "span",
-            marks: [],
-            text: "This is the text content for Carbon Accounting Explained 16.",
-          },
-        ],
-        style: "normal",
-      },
-    ],
-    date: "2024-07-24T09:09:02.000Z",
-  },
-];
